Store energy values as double instead of float
ev-cli: use double instead of float in generated code


libocpp: use double instead of float in generated code


libocpp: regenerate

Regenerated with the number->double generator change in place. Restricts
the diff to the float->double field changes in ocpp_types and a few
message types.


update hand-written parts of the code

everest_api_types: use double for power/energy API types

Part of the
float->double refactor. Converts the hand-written external API structs
(powermeter, energy, power_supply_DC, dc_external_derate, ocpp) to
double. Wrappers/codecs are cast-free and follow automatically;
transported as JSON over MQTT so no wire-format change.


libocpp: use double in hand-written smart charging logic

Aligns the hand-written OCPP code with the now-double generated types.
Converts sentinel constants (NO_SETPOINT_SPECIFIED / NO_DISCHARGE_LIMIT_
SPECIFIED switched to numeric_limits<double>::max()), PeriodLimit /
IntermediatePeriod / period_entry_t fields, profile-merging helpers, and
the non_finite<float> NaN/inf guard specializations.


iso15118/EvseV2G: use double for decoded physical values

Voltage/current/power/energy/SoC values decoded from EXI integers
(int16 x 10^multiplier) are internal session state with no wire
constraint. Converts DcMaximumLimits, the pre-charge callback, the V2G
session/config structs, and the iso/din server decode locals to double.
EXI codec types (cbv2g) stay integer-based; external generated struct
fields keep their explicit float casts.


libocpp: convert remaining hand-written code to double

Completes the float->double conversion across the OCPP library's
hand-written headers/sources that mirror or operate on the now-double
generated types: common/v2/v16 types.hpp, utils, evse, transaction,
database_handler, dynamic_schedule_manager, der_control, monitoring,
tariff_and_cost, device_model(_storage), init_device_model_db, and the
v16 charge_point. The library has no external wire-float types, so every
numeric value is double.


everest_api_types tests: teach autogenerator about double

The serialisation-test autogenerator only recognised float as a base
type; after the API structs moved to double it raised
NotImplementedError at configure time (breaking add_subdirectory on the
missing generated dir). Add double to base_types and the
value-generation switch.


libocpp tests: update mocks and tests to double

Match the test mocks, helpers, and value literals to the now-double
production types: the DatabaseHandlerMock insert_or_update_der_control
signature, smart-charging test utils/matchers, and the v16/v2/v21
profile and smart-charging tests (numeric_limits, optional, literals).


double conversions at module boundaries

Fixes the module-side fallout of the double switch: the iso15118
pre-charge callback definition, Evse15118D20 convert_from_optional (EXI
RationalNumber boundary now takes double), EvseManager AC power/voltage
locals feeding std::min against now-double capabilities, and the
LemDCBM/Isabellenhuette powermeter tests (FloatEq -> DoubleEq).


EnergyManager/EnergyNode: use double internally

The energy-trading core (Market, Broker, BrokerFastCharging) and the
EnergyNode schedule helpers still computed in float, leaking float
precision into the now-double energy types (e.g. ac_max_current_A =
power/voltage). Convert them to double for clean double-precision
output. Updates energy_node_tests expected values and two JSON fixtures
(5_ac_phase_switching, 9_setpoint_external_limit_watt) whose computed
currents are now full double precision (precision-only change).

+++ lib/everest/everest_api_types/include/everest_api_types/dc_external_derate/API.hpp
@@ -7,10 +7,10 @@
 namespace everest::lib::API::V1_0::types::dc_external_derate {
 
 struct ExternalDerating {
-    std::optional<float> max_export_current_A;
-    std::optional<float> max_import_current_A;
-    std::optional<float> max_export_power_W;
-    std::optional<float> max_import_power_W;
+    std::optional<double> max_export_current_A;
+    std::optional<double> max_import_current_A;
+    std::optional<double> max_export_power_W;
+    std::optional<double> max_import_power_W;
 };
 
 } // namespace everest::lib::API::V1_0::types::dc_external_derate
